use env params for PORT and DIR

+++ pkg/webserver/server.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 )
 
-func Serve(dir string, port uint) error {
+func Serve(dir string, port uint16) error {
 	fs := fileserver.New(fileserver.Options{})
 	http.Handle("/", fs.Serve(http.Dir(dir)))
 	return http.ListenAndServe(fmt.Sprintf(":%d", port), nil)
